Add "users" query
Queries can ask for `items` or `users`.  A side-effect is that you can have
empty queries if neither is specified.

Populate `first_name` and `last_name` in tests, so we can verify (some of)
the `fullName` logic.  It doesn't test cases where either `first_name` or
`last_name` is missing because the string concatenation logic whould have
to be duplicated between `UserType` and the spec.  I could have moved the
`full_name` method to the model, but this would obscure the demonstration
of custom attributes in GraphQL queries.